Move input mode to a more sensible place
