+++ src/InputGuide.tsx
@@ -1,7 +1,7 @@
 import {signal} from '@preact/signals';
 import type {FunctionComponent} from 'preact';
 import * as css from './InputGuide.module.css';
-import {activeInputMode, type InputMode} from './page-state/input-mode.js';
+import {activeInputMode, type InputMode} from './gamepad/input-mode.js';
 
 export type InputGuideIcon = {
 	readonly Icon: FunctionComponent;
